+++ app/src/main/java/com/example/braini/presentation/navigation/Screen.kt
@@ -1,0 +1,5 @@
+package com.example.braini.presentation.navigation
+
+sealed class Screen(val route: String) {
+    object Splash : Screen("screen_splash")
+}
